# BDN benchmark for custom operations.