--- a/benchmark/BDN.benchmark/Operations/CustomOperations.cs
+++ b/benchmark/BDN.benchmark/Operations/CustomOperations.cs
@@ -14,16 +14,27 @@ namespace BDN.benchmark.Operations
     [MemoryDiagnoser]
     public unsafe class CustomOperations : OperationsBase
     {
+        static ReadOnlySpan<byte> SETIFPM => "*4\r\n$7\r\nSETIFPM\r\n$1\r\nk\r\n$3\r\nval\r\n$1\r\nv\r\n"u8;
+        byte[] setIfPmRequestBuffer;
+        byte* setIfPmRequestBufferPointer;
+
         static ReadOnlySpan<byte> MYDICTSETGET => "*4\r\n$9\r\nMYDICTSET\r\n$2\r\nck\r\n$1\r\nf\r\n$1\r\nv\r\n*3\r\n$9\r\nMYDICTGET\r\n$2\r\nck\r\n$1\r\nf\r\n"u8;
         byte[] myDictSetGetRequestBuffer;
         byte* myDictSetGetRequestBufferPointer;
 
-        static ReadOnlySpan<byte> CPBSET => "*9\r\n$6\r\nCPBSET\r\n$6\r\n{0}000\r\n$6\r\n{0}001\r\n$6\r\n{0}002\r\n$6\r\n{0}003\r\n$6\r\n{0}000\r\n$6\r\n{0}001\r\n$6\r\n{0}002\r\n$6\r\n{0}003\r\n"u8;
-        byte[] cpbsetBuffer;
-        byte* cpbsetBufferPointer;
+        static ReadOnlySpan<byte> CTXNSET => "*9\r\n$7\r\nCTXNSET\r\n$6\r\n{0}000\r\n$6\r\n{0}001\r\n$6\r\n{0}002\r\n$6\r\n{0}003\r\n$6\r\n{0}000\r\n$6\r\n{0}001\r\n$6\r\n{0}002\r\n$6\r\n{0}003\r\n"u8;
+        byte[] ctxnsetBuffer;
+        byte* ctxnsetBufferPointer;
+
+        static ReadOnlySpan<byte> CPROCSET => "*9\r\n$8\r\nCPROCSET\r\n$6\r\n{0}000\r\n$6\r\n{0}001\r\n$6\r\n{0}002\r\n$6\r\n{0}003\r\n$6\r\n{0}000\r\n$6\r\n{0}001\r\n$6\r\n{0}002\r\n$6\r\n{0}003\r\n"u8;
+        byte[] cprocsetBuffer;
+        byte* cprocsetBufferPointer;
 
         void CreateExtensions()
         {
+            // Register custom raw string command
+            server.Register.NewCommand("SETIFPM", CommandType.ReadModifyWrite, new SetIfPMCustomCommand(), new RespCommandsInfo { Arity = 4 });
+
             // Register custom object type and commands
             var factory = new MyDictFactory();
             server.Register.NewType(factory);
@@ -31,29 +42,52 @@ namespace BDN.benchmark.Operations
             server.Register.NewCommand("MYDICTGET", CommandType.Read, factory, new MyDictGet(), new RespCommandsInfo { Arity = 3 });
 
             // Register custom transaction
-            server.Register.NewTransactionProc(CustomProcs.CustomProcSet.CommandName, () => new CustomProcSet(), new RespCommandsInfo { Arity = CustomProcs.CustomProcSet.Arity });
+            server.Register.NewTransactionProc(CustomProcs.CustomTxnSet.CommandName, () => new CustomTxnSet(),
+                new RespCommandsInfo { Arity = CustomProcs.CustomTxnSet.Arity });
+
+            // Register custom procedure
+            server.Register.NewProcedure(CustomProcs.CustomProcSet.CommandName, new CustomProcSet(),
+                new RespCommandsInfo { Arity = CustomProcs.CustomProcSet.Arity });
         }
+
         public override void GlobalSetup()
         {
             base.GlobalSetup();
             CreateExtensions();
 
+            SetupOperation(ref setIfPmRequestBuffer, ref setIfPmRequestBufferPointer, SETIFPM);
             SetupOperation(ref myDictSetGetRequestBuffer, ref myDictSetGetRequestBufferPointer, MYDICTSETGET);
-            SetupOperation(ref cpbsetBuffer, ref cpbsetBufferPointer, CPBSET);
+            SetupOperation(ref ctxnsetBuffer, ref ctxnsetBufferPointer, CTXNSET);
+            SetupOperation(ref cprocsetBuffer, ref cprocsetBufferPointer, CPROCSET);
+
+            SlowConsumeMessage("*4\r\n$7\r\nSETIFPM\r\n$1\r\nk\r\n$3\r\nval\r\n$1\r\nv\r\n"u8);
             SlowConsumeMessage("*4\r\n$9\r\nMYDICTSET\r\n$2\r\nck\r\n$1\r\nf\r\n$1\r\nv\r\n"u8);
-            SlowConsumeMessage(cpbsetBuffer);
+            SlowConsumeMessage(ctxnsetBuffer);
+            SlowConsumeMessage(cprocsetBuffer);
         }
 
         [Benchmark]
-        public void MyDictSetGet()
+        public void CustomRawStringCommand()
+        {
+            _ = session.TryConsumeMessages(setIfPmRequestBufferPointer, setIfPmRequestBuffer.Length);
+        }
+
+        [Benchmark]
+        public void CustomObjectCommand()
         {
             _ = session.TryConsumeMessages(myDictSetGetRequestBufferPointer, myDictSetGetRequestBuffer.Length);
         }
 
         [Benchmark]
-        public void CustomProcSet()
+        public void CustomTransaction()
         {
-            _ = session.TryConsumeMessages(cpbsetBufferPointer, cpbsetBuffer.Length);
+            _ = session.TryConsumeMessages(ctxnsetBufferPointer, ctxnsetBuffer.Length);
+        }
+
+        [Benchmark]
+        public void CustomProcedure()
+        {
+            _ = session.TryConsumeMessages(cprocsetBufferPointer, cprocsetBuffer.Length);
         }
     }
 }