--- a/benchmark/BDN.benchmark/Operations/ObjectOperations.cs
+++ b/benchmark/BDN.benchmark/Operations/ObjectOperations.cs
@@ -1,10 +1,7 @@
 ﻿// Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-using BDN.benchmark.CustomProcs;
 using BenchmarkDotNet.Attributes;
-using Garnet;
-using Garnet.server;
 
 namespace BDN.benchmark.Operations
 {
@@ -30,41 +27,19 @@ namespace BDN.benchmark.Operations
         byte[] hSetDelRequestBuffer;
         byte* hSetDelRequestBufferPointer;
 
-        static ReadOnlySpan<byte> MYDICTSETGET => "*4\r\n$9\r\nMYDICTSET\r\n$2\r\nck\r\n$1\r\nf\r\n$1\r\nv\r\n*3\r\n$9\r\nMYDICTGET\r\n$2\r\nck\r\n$1\r\nf\r\n"u8;
-        byte[] myDictSetGetRequestBuffer;
-        byte* myDictSetGetRequestBufferPointer;
-
-        static ReadOnlySpan<byte> CPBSET => "*9\r\n$6\r\nCPBSET\r\n$6\r\n{0}000\r\n$6\r\n{0}001\r\n$6\r\n{0}002\r\n$6\r\n{0}003\r\n$6\r\n{0}000\r\n$6\r\n{0}001\r\n$6\r\n{0}002\r\n$6\r\n{0}003\r\n"u8;
-        byte[] cpbsetBuffer;
-        byte* cpbsetBufferPointer;
-
-        void CreateExtensions()
-        {
-            var factory = new MyDictFactory();
-            server.Register.NewType(factory);
-            server.Register.NewCommand("MYDICTSET", CommandType.ReadModifyWrite, factory, new MyDictSet(), new RespCommandsInfo { Arity = 4 });
-            server.Register.NewCommand("MYDICTGET", CommandType.Read, factory, new MyDictGet(), new RespCommandsInfo { Arity = 3 });
-            server.Register.NewTransactionProc(CustomProcs.CustomProcSet.CommandName, () => new CustomProcSet(), new RespCommandsInfo { Arity = CustomProcs.CustomProcSet.Arity });
-        }
-
         public override void GlobalSetup()
         {
             base.GlobalSetup();
-            CreateExtensions();
             SetupOperation(ref zAddRemRequestBuffer, ref zAddRemRequestBufferPointer, ZADDREM);
             SetupOperation(ref lPushPopRequestBuffer, ref lPushPopRequestBufferPointer, LPUSHPOP);
             SetupOperation(ref sAddRemRequestBuffer, ref sAddRemRequestBufferPointer, SADDREM);
             SetupOperation(ref hSetDelRequestBuffer, ref hSetDelRequestBufferPointer, HSETDEL);
-            SetupOperation(ref myDictSetGetRequestBuffer, ref myDictSetGetRequestBufferPointer, MYDICTSETGET);
-            SetupOperation(ref cpbsetBuffer, ref cpbsetBufferPointer, CPBSET);
 
             // Pre-populate data
             SlowConsumeMessage("*4\r\n$4\r\nZADD\r\n$1\r\nc\r\n$1\r\n1\r\n$1\r\nd\r\n"u8);
             SlowConsumeMessage("*3\r\n$5\r\nLPUSH\r\n$1\r\nd\r\n$1\r\nf\r\n"u8);
             SlowConsumeMessage("*3\r\n$4\r\nSADD\r\n$1\r\ne\r\n$1\r\nb\r\n"u8);
             SlowConsumeMessage("*3\r\n$4\r\nHSET\r\n$1\r\nf\r\n$1\r\nb\r\n$1\r\nb\r\n"u8);
-            SlowConsumeMessage("*4\r\n$9\r\nMYDICTSET\r\n$2\r\nck\r\n$1\r\nf\r\n$1\r\nv\r\n"u8);
-            SlowConsumeMessage(cpbsetBuffer);
         }
 
         [Benchmark]
@@ -89,18 +64,6 @@ namespace BDN.benchmark.Operations
         public void HSetDel()
         {
             _ = session.TryConsumeMessages(hSetDelRequestBufferPointer, hSetDelRequestBuffer.Length);
-        }
-
-        [Benchmark]
-        public void MyDictSetGet()
-        {
-            _ = session.TryConsumeMessages(myDictSetGetRequestBufferPointer, myDictSetGetRequestBuffer.Length);
-        }
-
-        [Benchmark]
-        public void CustomProcSet()
-        {
-            _ = session.TryConsumeMessages(cpbsetBufferPointer, cpbsetBuffer.Length);
         }
     }
 }