Benchmark for all custom operations

- [x] Added custom procedure benchmark
- [x] Added custom raw string command benchmark
- [x] Moved custom object command and custom transaction from ObjectOperations to CustomOperations

main

| Method        | Params | Mean     | Error    | StdDev   | Allocated |
|-------------- |------- |---------:|---------:|---------:|----------:|
| CustomProcSet | None   | 70.30 us | 1.117 us | 0.933 us |         - |


| Method       | Params | Mean     | Error   | StdDev  | Gen0   | Allocated |
|------------- |------- |---------:|--------:|--------:|-------:|----------:|
| MyDictSetGet | None   | 120.8 us | 2.39 us | 2.94 us | 0.2441 |  23.44 KB |

main's CustomProcSet maps to PR's CustomTransaction
main's MyDictSetGet maps to PR's CustomObjectCommand

PR 

| Method                 | Params | Mean      | Error    | StdDev   | Median    | Gen0   | Allocated |
|----------------------- |------- |----------:|---------:|---------:|----------:|-------:|----------:|
| CustomRawStringCommand | None   |  30.80 us | 0.484 us | 0.429 us |  30.82 us |      - |         - |
| CustomObjectCommand    | None   | 124.61 us | 2.441 us | 3.654 us | 124.42 us | 0.2441 |   24000 B |
| CustomTransaction      | None   |  73.31 us | 1.456 us | 1.495 us |  73.54 us |      - |         - |
| CustomProcedure        | None   |  57.40 us | 1.127 us | 1.206 us |  57.22 us |      - |         - |